+++ src/pandoc.jl
@@ -5,15 +5,13 @@
 
 Convert output from pandoc markdown to html using Weave.jl template
 """
-function pandoc2html(formatted::AbstractString, doc::WeaveDoc)
+function pandoc2html(formatted::AbstractString, doc::WeaveDoc, outname::AbstractString)
   html_template = joinpath(Pkg.dir("Weave"), "templates/pandoc_skeleton.html")
   css_template = joinpath(Pkg.dir("Weave"), "templates/pandoc_skeleton.css")
 
   path, wsource = splitdir(abspath(doc.source))
   wversion = string(Pkg.installed("Weave"))
   wtime =  Date(now())
-
-  outname = "$(doc.basename).html"
 
   #Change path for pandoc
   old_wd = pwd()
@@ -40,7 +38,7 @@ end
 
 Convert output from pandoc markdown to pdf using Weave.jl template
 """
-function pandoc2pdf(formatted::AbstractString, doc::WeaveDoc)
+function pandoc2pdf(formatted::AbstractString, doc::WeaveDoc, outname::AbstractString)
 
   header_template = joinpath(Pkg.dir("Weave"), "templates/pandoc_header.txt")
 
@@ -48,7 +46,6 @@ function pandoc2pdf(formatted::AbstractString, doc::WeaveDoc)
   wversion = string(Pkg.installed("Weave"))
   wtime =  Date(now())
 
-  outname = "$(doc.basename).pdf"
 
   #Change path for pandoc
   old_wd = pwd()