+++ test/chunk_options.jl
@@ -28,11 +28,28 @@ ref = @compat readstring(open("documents/chunk_options_ref.rst"))
 @test result == ref
 cleanup && rm("documents/chunk_options.rst")
 
+#Test out_path
+weave("documents/chunk_options.noweb", doctype="rst",
+      out_path="documents/outpath_options.rst" , plotlib=nothing)
+result = @compat readstring(open("documents/outpath_options.rst"))
+ref = @compat readstring(open("documents/chunk_options_ref.rst"))
+@test result == ref
+cleanup && rm("documents/outpath_options.rst")
+
+#Test tangle
 tangle("documents/chunk_options.noweb")
 result = @compat readstring(open("documents/chunk_options.jl"))
 ref = @compat readstring(open("documents/chunk_options_ref.jl"))
 @test result == ref
 cleanup && rm("documents/chunk_options.jl")
+
+
+tangle("documents/chunk_options.noweb", out_path = "documents/outoptions.jl")
+result = @compat readstring(open("documents/outoptions.jl"))
+ref = @compat readstring(open("documents/chunk_options_ref.jl"))
+@test result == ref
+cleanup && rm("documents/outoptions.jl")
+
 
 #Test functions and sandbox clearing
 weave("documents/chunk_func.noweb", plotlib=nothing)